requirements.txt
numpy==1.18.5
uvicorn==0.15.0
opencv-python-headless
fastapi==0.68.1
Pillow==8.3.2
python-multipart
